checkpointArchives {checkpoint}	R Documentation

List checkpoint archives on disk.

Description

List checkpoint archives on disk.

Usage

checkpointArchives(checkpointLocation = "~/", full.names = FALSE)

Arguments

checkpointLocation
File path where the checkpoint library is stored. Default is "~/", i.e. the user's home directory. A use case for changing this is to create a checkpoint library on a portable drive (e.g. USB drive).

full.names
passed to list.files()

See Also

Other checkpoint functions: checkpointRemove, checkpoint, getAccessDate, getValidSnapshots, mranUrl, setSnapshot, unCheckpoint

Examples

checkpointArchives()
## Not run: 
checkpointRemove("2016-10-01")

## End(Not run)

[Package checkpoint version 0.4.9 Index]