checkpointRemove {checkpoint}	R Documentation

Remove checkpoint archive from disk.

Description

This function enables you to delete a snapshot archive folder from disk, thus releasing storage space. If you supply a single snapshotDate, then only this archive will be removed. You also have the option to remove a series of snapshots, including all snapshots before a given date, or all snapshots that have not been accessed since a given date.

Usage

checkpointRemove(snapshotDate, checkpointLocation = "~/",
  allSinceSnapshot = FALSE, allUntilSnapshot = FALSE,
  notUsedSince = FALSE)

Arguments

snapshotDate
Date of snapshot to use in YYYY-MM-DD format, e.g. "2014-09-17". Specify a date on or after "2014-09-17". MRAN takes one snapshot per day. To list all valid snapshot dates on MRAN use getValidSnapshots()

checkpointLocation
File path where the checkpoint library is stored. Default is "~/", i.e. the user's home directory. A use case for changing this is to create a checkpoint library on a portable drive (e.g. USB drive).

allSinceSnapshot
If TRUE, removes all snapshot archives since the snapshotDate

allUntilSnapshot
If TRUE, removes all snapshot archives before the snapshotDate

notUsedSince
If TRUE, removes all snapshot archives that have not been accessed since the snapshotDate. See getAccessDate()

See Also

getAccessDate()

Other checkpoint functions: checkpointArchives, checkpoint, getAccessDate, getValidSnapshots, mranUrl, setSnapshot, unCheckpoint

Examples

checkpointArchives()
## Not run: 
checkpointRemove("2016-10-01")

## End(Not run)

[Package checkpoint version 0.4.3 Index]