getAccessDate {checkpoint}	R Documentation

Returns the date the snapshot was last accessed.

Description

The checkpoint() function stores a marker in the snapshot folder every time the function gets called. This marker contains the system date, thus indicating the the last time the snapshot was accessed.

Usage

getAccessDate(checkpointLocation = "~/")

Arguments

checkpointLocation
File path where the checkpoint library is stored. Default is "~/", i.e. the user's home directory. A use case for changing this is to create a checkpoint library on a portable drive (e.g. USB drive).

Value

Named character with last access date

See Also

checkpointRemove()

Other checkpoint functions: checkpointArchives, checkpointRemove, checkpoint, getValidSnapshots, mranUrl, setSnapshot, unCheckpoint

[Package checkpoint version 0.4.3 Index]