scanForPackages {checkpoint}	R Documentation

Scans a project (or folder) for references to packages.

Description

Scans a project (or folder) for references to packages.

Usage

scanForPackages(project = getwd(), verbose = TRUE, use.knitr = FALSE,
  auto.install.knitr = FALSE, scan.rnw.with.knitr = FALSE)

Arguments

project
A project path. This is the path to the root of the project that references the packages to be installed from the MRAN snapshot for the date specified for snapshotDate. Defaults to current working directory using getwd().

verbose
If TRUE, displays progress messages.

use.knitr
If TRUE, parses all Rmarkdown files using the knitr package.

auto.install.knitr
If TRUE and the project contains rmarkdown files, then automatically included the packages knitr in packages to install.

scan.rnw.with.knitr
If TRUE, uses knitr::knit() to parse .Rnw files, otherwise use utils::Sweave()

Value

A list with two elements:

pkgs: a character vector with the names of identified packages

error: a character vector with information about files that could not be parsed

[Package checkpoint version 0.4.3 Index]