unCheckpoint {checkpoint}	R Documentation

Undo the effect of checkpoint by resetting .libPath to user library location.

Description

This is an experimental solution to the situation where a user no longer wants to work in the checkpointed environment. The function resets .libPaths to its pre-checkpoint value.

Note that this does not undo any of the other side-effects of checkpoint(). Specifically, all loaded packages remain loaded, and the value of getOption("repos") remains unchanged.

Usage

unCheckpoint(new)

Arguments

new
Not used; for back-compatibility only.

See Also

Other checkpoint functions: checkpointArchives, checkpointRemove, checkpoint, getAccessDate, getValidSnapshots, mranUrl, setSnapshot

[Package checkpoint version 0.4.9 Index]